re-factor: move createFetch, getProxyOptionsFromConfig to utils
